[DDC-2387] Fix DatabaseDriver not working with combinations of composite/association keys.
