+++ lib/Doctrine/ORM/Mapping/Driver/DatabaseDriver.php
@@ -217,7 +217,8 @@ class DatabaseDriver implements MappingDriver
         }
 
         if ($ids) {
-            if (count($ids) == 1) {
+            // We need to check for the columns here, because we might have associations as id as well.
+            if (count($primaryKeyColumns) == 1) {
                 $metadata->setIdGeneratorType(ClassMetadataInfo::GENERATOR_TYPE_AUTO);
             }
