+++ tests/Doctrine/Tests/ORM/Functional/DatabaseDriverTest.php
@@ -2,12 +2,9 @@
 
 namespace Doctrine\Tests\ORM\Functional;
 
-require_once __DIR__ . '/../../TestInit.php';
+use Doctrine\ORM\Mapping\ClassMetadataInfo;
 
-use Doctrine\ORM\Mapping\ClassMetadataInfo,
-    Doctrine\Common\Util\Inflector;
-
-class DatabaseDriverTest extends \Doctrine\Tests\OrmFunctionalTestCase
+class DatabaseDriverTest extends DatabaseDriverTestCase
 {
     /**
      * @var \Doctrine\DBAL\Schema\AbstractSchemaManager
@@ -147,45 +144,5 @@ class DatabaseDriverTest extends \Doctrine\Tests\OrmFunctionalTestCase
         $metadatas = $this->convertToClassMetadata(array($tableA, $tableB), array($tableMany));
 
         $this->assertEquals(0, count($metadatas['DbdriverBaz']->associationMappings), "no association mappings should be detected.");
-    }
-
-    protected function convertToClassMetadata(array $entityTables, array $manyTables = array())
-    {
-        $driver = new \Doctrine\ORM\Mapping\Driver\DatabaseDriver($this->_sm);
-        $driver->setTables($entityTables, $manyTables);
-
-        $metadatas = array();
-        foreach ($driver->getAllClassNames() AS $className) {
-            $class = new ClassMetadataInfo($className);
-            $driver->loadMetadataForClass($className, $class);
-            $metadatas[$className] = $class;
-        }
-
-        return $metadatas;
-    }
-
-    /**
-     * @param  string $className
-     * @return ClassMetadata
-     */
-    protected function extractClassMetadata(array $classNames)
-    {
-        $classNames = array_map('strtolower', $classNames);
-        $metadatas = array();
-
-        $driver = new \Doctrine\ORM\Mapping\Driver\DatabaseDriver($this->_sm);
-        foreach ($driver->getAllClassNames() as $className) {
-            if (!in_array(strtolower($className), $classNames)) {
-                continue;
-            }
-            $class = new ClassMetadataInfo($className);
-            $driver->loadMetadataForClass($className, $class);
-            $metadatas[$className] = $class;
-        }
-
-        if (count($metadatas) != count($classNames)) {
-            $this->fail("Have not found all classes matching the names '" . implode(", ", $classNames) . "' only tables " . implode(", ", array_keys($metadatas)));
-        }
-        return $metadatas;
     }
 }
